server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "The Pet Headquarters — Backend API"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
    "python-decouple>=3.8",
    "pyjwt>=2.9",
    "stripe>=12.0",
    "cloudinary>=1.40",
    "pillow>=11.0",
]
